Add VectorBase parent class templates for Vector...
This system of VectorBase specializations allow for mixing in members
for a given index of the vector which will be inherited by all vectors
having size enough for that index.  So for example, the one for size 1
introduces an 'x' member and the one for size 2 introduces 'y', and so
forth, and any Vector that is large enough will contain all of them.